Import zfs-mirror
Currently this script will check all datasets on a specified pool, and
where it finds a local property specifying a destination host and pool,
will send a snapshot of that dataset to the destination. Called
repeatedly, it will send incremental snapshots, storing the most recent
as a zfs property on the source.

It is safe to call as often as desired, even if a previous instance is
already running - in that case it will keep trying for a while, then
give up and do nothing if it can't get a lock.