Take snapshots individually if necessary
Some zfs versions don't allow multiple snapshots to be made in a single
invocation. Test if this is the case (currently the only means I know of
is scraping the usage note), and create all necessary snapshots in
separate invocations if necessary.

This will produce a warning, since snapshot creation is not an atomic
operation in this case.