Use ": $((arithmetic))" rather than "((arithmetic))"
Since bash 4.1, arithmetic expressions of the form ((something)) will
trigger 'set -e' to abort if the result of the expression is zero.

This was causing the script to abort whenever the result of expressions
like ((DEPTH++)) was zero. In this specific case, using ((++DEPTH))
would have worked around it, but : $((DEPTH++)) is the standard way to
ignore the value of the expression, so is the better solution even
though it looks a little unusual.