Improve output of age reporting in error cases
The existence of the source and destination pools is tested before
running the check, as otherwise specifying an invalid pool would cause
no results to be printed, as if everything is up to date.

Also, when '-v' is specified, be explicit about what pool is being
checked, and display a notification if there appears to be a
synchronisation currently in progress to the given destination.