Add ability to specify multiple destinations
The destination argument now accepts a comma-separated list.

Several errors are demoted to warnings, where it would be possible to
continue to do some useful work. This is intended to avoid the
situation where two pools are given, but an error with the first
prevents the second from even being tried.

If any warnings are encountered which would previously have been
errors, the return value of the script is the most recent error code.