Allow dest to be specified even if src is omitted
Previously, if the source pool argument was omitted, the destination
also had to be omitted. Now, if there is only one pool non-option
argument, the script checks if it looks like a source or a destination
by looking for a colon, which signifies a destination (of the form
machine:pool).