Make dest lock per-source
Technically there's no reason that simultaneous transfers from
different source pools to the same destination should be disallowed.
This is equivalent to mirroring to a given destination from two
different machines, and will only break if there is a dataset naming
conflict, in which case the conflict will exist regardless of whether
the transfers are concurrent or sequential.

Hence, this relaxes the restriction and allows runs to be made
concurrently to a given destination, so long as the source pool is
different. The effective result is that it's possible, for example, to
keep periodically synchronising from one pool even if there is a long
ongoing transfer from another pool that we would otherwise have to wait
for.