Add option to wait for source/dest pool locks
Normally if there's a sync in progress we for a given
source/destination pair we skip it, in order to allow the script to be
called arbitrarily often without worrying about building up a backlog
of waiting processes.

With -w, it is now possible to wait indefinitely until the pools are
available. This is mostly useful for queuing up a sync using different
options (like -v and -p) rather than having to kill a sync in progress
or manually wait until it's done.

Since the process only polls for the lockfile availability every ten
seconds, it is possible for another process to jump the queue, however
in normal circumstances if you're not running the script from cron
every few seconds, the odds of being the first to get the lock are
pretty good.