Don't implicitly try invalid src/dest combinations
Previously when a user requested a destination but no source, we would
try sending to that destination from *all* potential sources, producing
a warning when an implicitly created source/destination pair was
invalid.

If a user *explicitly* requested an invalid source/destination pair,
this probably represents a mistake, and so we do want to check and
report these. On the other hand, if a user simply selects a destination,
then we need to determine the sources that are configured to actually
use it.