Improve zfs_send_new when dest already exists
When zfs_send_new is called, this should be the first time the
dataset has been synchronised, but it's possible to be called in the
case that the dataset exists but there is no zfs-mirror snapshot on it
- for example, if it's been sent previously via some other means, or if
the zfs-mirror snapshots have been destroyed for some reason.

If the destination has any snapshots, and if the source has a snapshot
corresponding to the destination's most recent, then we attempt to
proceed by sending an incremental stream to get back on track. If not,
we can at least produce a meaningful message rather than just failing.