Stop using a pool-global 'latest' property
Rather than using a property to store the latest snapshot number that
this pool should be using, use whatever is the latest snapshot that we
actually have.

This allows each dataset's latest snapshot to be independent of the
others, which among other things makes it easier to move datasets
between pools without having to rename all their snapshots.